Rename references.ts to helpers.ts.
