+++ packages/apollo-cache-inmemory/src/index.ts
@@ -8,4 +8,8 @@ export * from './readFromStore';
 export * from './writeToStore';
 export * from './objectCache';
 export * from './types';
-export * from './references';
+export {
+  Reference,
+  makeReference,
+  isReference,
+} from './helpers';
